Added support for private tasks (per user)

+++ src/tasks/tasks.service.ts
@@ -7,6 +7,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -16,33 +17,42 @@ export class TasksService {
 
     ) {}
     
-    getTasks(filterDto: GetTasksFilterDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto); 
+    getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User,
+        ):Promise<Task[]>{
+        return this.taskRepository.getTasks(filterDto, user); 
     }
 
-    async getTaskById(id: number): Promise<Task>{
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task>{
+        
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id} });
         if(!found){
             throw new NotFoundException(`Task with id "${id}" not found`);
         }
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-        return this.taskRepository.createTask(createTaskDto)
+    async createTask(createTaskDto: CreateTaskDto,
+        user: User): Promise<Task>{
+        return this.taskRepository.createTask(createTaskDto, user)
     }
 
-    async deleteTaskById(id: number): Promise<void> {
+    async deleteTaskById(id: number, user: User): Promise<void> {
         
-        const result = await this.taskRepository.delete(id);
+        const result = await this.taskRepository.delete({ id, userId: user.id});
 
         if(result.affected === 0) {
             throw new NotFoundException(`Task with id "${id}" not found`);
         }
         
     }
-    async updateTaskStatus(id: number, status: TaskStatus):Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User,
+        ):Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
